extract item styles into a file
